refactor(core): extract rls filter staging helpers into rls-utils
Mirrors upsert-utils: findRlsFilterDefs (with its MetadataStorage-keyed
cache), computeRlsFilterVariables, the prune computation, and the cache
invalidation hook move out of EntityManager into free functions.
